phong model is working (tests 1 and 2)

+++ src/raytracer/pigments/Finish.java
@@ -1,9 +1,9 @@
 package raytracer.pigments;
 
 public class Finish {
-	public double amb, diff, spec, shiny, refl, trans, ior;
+	public float amb, diff, spec, shiny, refl, trans, ior;
 
-	public Finish(double amb, double diff, double spec, double shiny, double refl, double trans, double ior) {
+	public Finish(float amb, float diff, float spec, float shiny, float refl, float trans, float ior) {
 		this.amb = amb;
 		this.diff = diff;
 		this.spec = spec;
